requirements.txt
google-cloud-aiplatform==1.35.0
langchain==0.0.323
faiss-cpu==1.7.4
google-auth==2.29.0
langchain-google-vertexai==1.0.4